test(processing): move csg_summary tests to a sibling _tests.rs file
Inline #[cfg(test)] mod tests lived inside csg_summary.rs, one of the
production files check-test-revert-oracle.mjs reverts to prove PR #4081's
changed test observes the production change -- reverting a brand new file
deletes it and its tests together, so the oracle reported UNOBSERVED.
Same pattern as determinism_tests.rs / simplify_session_tests.rs.
